chore(dual-channel): renamed dev/data.json -> src/test-data/data.json
