+++ packages/dual-channel/src/test-data/data.json
